+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unpublished Changes
+* larva-patterns - Add `narrow` variant for `container` module.
 
 ## 1.26.0 09-01-2022
 * larva - Add repo-wide JS linting and apply auto-linting fixes.